Changed the schema updated profile and few small stuffs

+++ frontend/src/pages/SignUp.jsx
@@ -44,6 +44,7 @@ const Input = styled.input`
   border: 1px solid ${({ theme }) => theme.hr};
   border-radius: 3px;
   padding: 10px;
+  margin: "0 10px";
   background-color: transparent;
   color: ${({ theme }) => theme.text};
   width: 100%;
@@ -58,10 +59,29 @@ const Button = styled.button`
   background-color: ${({ theme }) => theme.navbar};
   color: ${({ theme }) => theme.text};
 `;
+const Select = styled.select`
+  border: 1px solid ${({ theme }) => theme.hr};
+  border-radius: 3px;
+  padding: 10px;
+  background-color: transparent;
+  color: ${({ theme }) => theme.text};
+  width: 111%;
+  box-sizing: border-box; /* Add this line to include padding and border in the width calculation */
+`;
 
+const Option = styled.option`
+  color: black;
+`;
 const SignUp = () => {
   const navigate = useNavigate();
-  const [user, setUser] = useState({ name: "", email: "", password: "" });
+  const [user, setUser] = useState({
+    name: "",
+    email: "",
+    password: "",
+    confirmPassword: "",
+    gender: "",
+    dob: "",
+  });
 
   const handleInputs = (e) => {
     console.log(e);
@@ -71,19 +91,26 @@ const SignUp = () => {
   const handleSignup = async (e) => {
     e.preventDefault();
     try {
-      const { name, email, password } = user;
-      const response = await axios.post("http://localhost:8000/signup", {
-        name,
-        email,
-        password,
-      });
-      if (response.status === 201) {
-        window.alert("Registration Successful");
-        localStorage.setItem("user", JSON.stringify(response));
-        console.log("Successful Registration");
-      } else {
-        window.alert("Invalid Registration");
-        console.log("Invalid Registration");
+      const { name, email, password, confirmPassword, gender, dob } = user;
+      if (password !== confirmPassword) {
+        window.alert("Passwords do not match");
+        return;
+      } else if (password === confirmPassword) {
+        const response = await axios.post("http://localhost:8000/signup", {
+          name,
+          email,
+          password,
+          gender,
+          dob,
+        });
+        if (response.status === 201) {
+          window.alert("Registration Successful");
+          localStorage.setItem("user", JSON.stringify(response));
+          console.log("Successful Registration");
+        } else {
+          window.alert("Invalid Registration");
+          console.log("Invalid Registration");
+        }
       }
     } catch (err) {
       // Handle error response
@@ -124,6 +151,27 @@ const SignUp = () => {
           value={user.password}
           onChange={handleInputs}
         />
+        <Input
+          type="password"
+          placeholder="Confirm Password"
+          name="confirmPassword"
+          value={user.confirmPassword}
+          onChange={handleInputs}
+        />
+        <Select name="gender" value={user.gender} onChange={handleInputs}>
+          <Option value="">Select Gender</Option>
+          <Option value="male">Male</Option>
+          <Option value="female">Female</Option>
+          <Option value="other">Other</Option>
+        </Select>
+        <Input
+          type="date"
+          placeholder="Date of Birth"
+          name="dob"
+          value={user.dob}
+          onChange={handleInputs}
+        />
+
         <Button className="signup-button" onClick={handleSignup}>
           Sign Up
         </Button>
